+++ onsenui/esm/elements/ons-list-item.js
@@ -173,7 +173,6 @@ export default class ListItemElement extends BaseElement {
     util.defineBooleanProperty(this, 'expanded');
 
     this._animatorFactory = this._updateAnimatorFactory();
-    this.toggleExpansion = this.toggleExpansion.bind(this);
 
     // Elements ignored when tapping
     const re = /^ons-(?!col$|row$|if$)/i;
@@ -336,8 +335,6 @@ export default class ListItemElement extends BaseElement {
 
     this._expanding = true;
 
-    this.dispatchEvent(new Event('expansion'));
-
     const animator = this._animatorFactory.newAnimator();
     animator._animateExpansion(this, this.expanded, () => {
       this._expanding = false;
@@ -416,8 +413,13 @@ export default class ListItemElement extends BaseElement {
     this[action]('mouseout', this._onRelease);
 
     if (this._top) {
-      this._top[action]('click', this.toggleExpansion);
+      this._top[action]('click', this._onClickTop.bind(this));
     }
+  }
+
+  _onClickTop() {
+    this.toggleExpansion();
+    this.dispatchEvent(new Event('expansion'));
   }
 
   _onDrag(event) {